Initial extraction of scalar rt math
